move global files to global folder
