+++ src/ensembl/src/Root.tsx
@@ -12,11 +12,11 @@ import Header from './header/Header';
 import Content from './content/Content';
 import PrivacyBanner from './shared/privacy-banner/PrivacyBanner';
 
-import { updateBreakpointWidth } from './globalActions';
-import { getBreakpointWidth } from './globalSelectors';
+import { updateBreakpointWidth } from './global/globalActions';
+import { getBreakpointWidth } from './global/globalSelectors';
 import { RootState } from './store';
-import { BreakpointWidth } from './globalConfig';
-import { getBreakpoint } from './globalHelper';
+import { BreakpointWidth } from './global/globalConfig';
+import { getBreakpoint } from './global/globalHelper';
 
 import styles from './Root.scss';
