+++ src/ensembl/src/global/globalActions.ts
@@ -1,7 +1,7 @@
 import { createAction } from 'typesafe-actions';
 
 import { BreakpointWidth } from './globalConfig';
-import { getGlobalAnalyticsObject } from './analyticsHelper';
+import { getGlobalAnalyticsObject } from '../analyticsHelper';
 
 export const updateBreakpointWidth = createAction(
   'browser/update-breakpoint-width',
