+++ src/ensembl/src/global/globalConfig.ts
