+++ src/ensembl/src/global/globalHelper.ts
