+++ src/ensembl/src/global/globalReducer.ts
