+++ src/ensembl/src/global/globalSelectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from './store';
+import { RootState } from '../store';
 import { BreakpointWidth } from './globalConfig';
 
 export const getBreakpointWidth = (state: RootState): BreakpointWidth =>
