+++ src/ensembl/src/global/globalState.ts
